feat: separate new widget and widget dialog
